backend/requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distlib==0.3.9
Django==5.1
dnspython==2.7.0
email_validator==2.2.0
filelock==3.17.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.10.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fpdf==1.7.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mongoengine==0.29.1
numpy==2.2.2
opencv-python==4.8.0.74
pandas==2.2.3
Pillow==10.0.0
platformdirs==4.3.6
PyJWT==2.10.1
pymongo==4.5.0
pypng==0.20220715.0
PyQt6==6.5.0
PyQt6-Qt6==6.5.0
PyQt6_sip==13.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
qrcode==7.4.2
requests==2.31.0
setuptools==72.2.0
six==1.17.0
SQLAlchemy==2.0.37
sqlparse==0.5.1
stripe==11.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
virtualenv==20.29.1
Werkzeug==3.1.3
WTForms==3.2.1
